requirements.txt
asgiref==3.7.2
Django==4.2.7
django-active-link==0.1.8
django-filter==23.4
sqlparse==0.4.4
tzdata==2023.3